moved check_profile_for_errors to Host.rb
